requirements.txt
flask==2.0.1
werkzeug==2.0.1
python-dotenv==1.0.0
manim
openai==1.3.5
requests==2.26.0
gunicorn==20.1.0
numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.0.0
pycairo>=1.21.0
ffmpeg-python>=0.2.0
moderngl>=5.6.4
moderngl-window>=2.4.1
watchdog>=2.1.3
opencv-python
langgraph
langchain-openai
moviepy
azure-cognitiveservices-speech
google-genai
pytest>=7.0.0
pytest>=7.0.0